+++ src/eternity.rs
@@ -125,9 +125,9 @@ impl Eternity {
             .item_index
             .get(&item)
             .expect("if item is present in block index, it must be present in item index");
-        let index = ((epoch_in_eternity as usize) << 32)
-            | ((block_in_epoch as usize) << 16)
-            | item_in_block as usize;
+        let index = ((epoch_in_eternity as u64) << 32)
+            | ((block_in_epoch as u64) << 16)
+            | item_in_block as u64;
 
         let (auth_path, leaf) = self.inner.witness(index)?;
         Some(Proof {
