+++ src/main.rs
@@ -3,7 +3,7 @@ use penumbra_tct::{Block, Epoch, Eternity, Fq, Hash, Insert, PrimeField};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut tree = Eternity::new();
-    for (i, keep) in (0usize..1000).zip([true].into_iter().cycle()) {
+    for (i, keep) in (0u64..1000).zip([true].into_iter().cycle()) {
         let fq = Fq::from_le_bytes_mod_order(&i.to_le_bytes());
         if keep {
             tree.insert_item(Insert::Keep(fq)).unwrap();
@@ -14,6 +14,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let root = tree.hash();
     let proof = tree.witness(0).unwrap();
     assert!(proof.verify(root).is_ok());
-    assert!(tree.witness(usize::MAX).is_none());
+    assert!(tree.witness(u64::MAX).is_none());
     Ok(())
 }
